API Library

API Library

Module

Base Kernels API

Composite Kernels

Transforms

Functions

Utilities

Index

Module

KernelFunctions.KernelFunctions — Module

KernelFunctions. Github Documentation

Base Kernels API

KernelFunctions.ConstantKernel — Type

ConstantKernel(; c=1.0)

Kernel function always returning a constant value c

    κ(x,y) = c

KernelFunctions.WhiteKernel — Type

WhiteKernel()

    κ(x,y) = δ(x,y)

Kernel function working as an equivalent to add white noise. Can also be called via EyeKernel()

KernelFunctions.EyeKernel — Type

EyeKernel()

See WhiteKernel

KernelFunctions.ZeroKernel — Type

ZeroKernel()

Create a kernel that always returns zero

    κ(x,y) = 0.0

The output type depends on x and y

KernelFunctions.CosineKernel — Type

CosineKernel()

The cosine kernel is a stationary kernel for a sinusoidal given by

    κ(x,y) = cos( π * (x-y) )

KernelFunctions.SqExponentialKernel — Type

SqExponentialKernel()

The squared exponential kernel is a Mercer kernel given by the formula:

    κ(x, y) = exp(-‖x - y‖² / 2)

Can also be called via RBFKernel, GaussianKernel or SEKernel. See GammaExponentialKernel for a generalization.

KernelFunctions.ExponentialKernel — Type

ExponentialKernel()

The exponential kernel is a Mercer kernel given by the formula:

    κ(x,y) = exp(-‖x-y‖)

Can also be called via LaplacianKernel or Matern12Kernel. See GammaExponentialKernel for a generalization.

KernelFunctions.GammaExponentialKernel — Type

GammaExponentialKernel(; γ = 2.0)

The γ-exponential kernel [1] is an isotropic Mercer kernel given by the formula:

    κ(x,y) = exp(-‖x-y‖^γ)

Where γ > 0, (the keyword γ can be replaced by gamma) For γ = 2, see SqExponentialKernel; for γ = 1, see ExponentialKernel.

[1] - Gaussian Processes for Machine Learning, Carl Edward Rasmussen and Christopher K. I. Williams, MIT Press, 2006.

KernelFunctions.ExponentiatedKernel — Type

ExponentiatedKernel()

The exponentiated kernel is a Mercer kernel given by:

    κ(x,y) = exp(xᵀy)

KernelFunctions.FBMKernel — Type

FBMKernel(; h::Real=0.5)

Fractional Brownian motion kernel with Hurst index h from (0,1) given by

    κ(x,y) =  ( |x|²ʰ + |y|²ʰ - |x-y|²ʰ ) / 2

For h=1/2, this is the Wiener Kernel, for h>1/2, the increments are positively correlated and for h<1/2 the increments are negatively correlated.

KernelFunctions.GaborKernel — Type

GaborKernel(; ell::Real=1.0, p::Real=1.0)

Gabor kernel with lengthscale ell and period p. Given by

\[ κ(x,y) = h(x-z), h(t) = exp(-sum(t.^2./(ell.^2)))*cos(pi*sum(t./p))\]

KernelFunctions.MaternKernel — Type

MaternKernel(; ν = 1.0)

The matern kernel is a Mercer kernel given by the formula:

    κ(x,y) = 2^{1-ν}/Γ(ν)*(√(2ν)‖x-y‖)^ν K_ν(√(2ν)‖x-y‖)

For ν=n+1/2, n=0,1,2,... it can be simplified and you should instead use ExponentialKernel for n=0, Matern32Kernel, for n=1, Matern52Kernel for n=2 and SqExponentialKernel for n=∞.

KernelFunctions.Matern32Kernel — Type

Matern32Kernel()

The matern 3/2 kernel is a Mercer kernel given by the formula:

    κ(x,y) = (1+√(3)‖x-y‖)exp(-√(3)‖x-y‖)

KernelFunctions.Matern52Kernel — Type

Matern52Kernel()

The matern 5/2 kernel is a Mercer kernel given by the formula:

    κ(x,y) = (1+√(5)‖x-y‖ + 5/3‖x-y‖^2)exp(-√(5)‖x-y‖)

KernelFunctions.NeuralNetworkKernel — Type

NeuralNetworkKernel()

Neural network kernel function.

\[ κ(x, y) = asin(x' * y / sqrt[(1 + x' * x) * (1 + y' * y)])\]

Significance

Neal (1996) pursued the limits of large models, and showed that a Bayesian neural network becomes a Gaussian process with a neural network kernel as the number of units approaches infinity. Here, we give the neural network kernel for single hidden layer Bayesian neural network with erf (Error Function) as activation function.

References:

GPML Pg 105

Neal(1996)

Andrew Gordon's Thesis Pg 45

KernelFunctions.LinearKernel — Type

LinearKernel(; c = 0.0)

The linear kernel is a Mercer kernel given by

    κ(x,y) = xᵀy + c

Where c is a real number

KernelFunctions.PolynomialKernel — Type

PolynomialKernel(; d = 2.0, c = 0.0)

The polynomial kernel is a Mercer kernel given by

    κ(x,y) = (xᵀy + c)^d

Where c is a real number, and d is a shape parameter bigger than 1. For d = 1 see LinearKernel

KernelFunctions.PiecewisePolynomialKernel — Type

PiecewisePolynomialKernel{V}(maha::AbstractMatrix)

Piecewise Polynomial covariance function with compact support, V = 0,1,2,3. The kernel functions are 2V times continuously differentiable and the corresponding processes are hence V times mean-square differentiable. The kernel function is:

\[ κ(x, y) = max(1 - r, 0)^(j + V) * f(r, j) with j = floor(D / 2) + V + 1\]

where r is the Mahalanobis distance mahalanobis(x,y) with maha as the metric.

KernelFunctions.MahalanobisKernel — Type

MahalanobisKernel(P::AbstractMatrix)

Mahalanobis distance-based kernel given by

\[ κ(x,y) = exp(-r^2), r^2 = maha(x,P,y) = (x-y)'* P *(x-y)\]

where the matrix P is the metric.

KernelFunctions.RationalQuadraticKernel — Type

RationalQuadraticKernel(; α=2.0)

The rational-quadratic kernel is a Mercer kernel given by the formula:

    κ(x, y) = (1 + ||x − y||² / (2α))^(-α)

where α is a shape parameter of the Euclidean distance. Check GammaRationalQuadraticKernel for a generalization.

KernelFunctions.GammaRationalQuadraticKernel — Type

GammaRationalQuadraticKernel([α=2.0 [, γ=2.0]])

The Gamma-rational-quadratic kernel is an isotropic Mercer kernel given by the formula:

    κ(x, y) = (1 + ||x−y||^γ / α)^(-α)

where α is a shape parameter of the Euclidean distance and γ is another shape parameter.

KernelFunctions.spectral_mixture_kernel — Function

spectral_mixture_kernel(

    h::Kernel=SqExponentialKernel(),
    αs::AbstractVector{<:Real},
    γs::AbstractMatrix{<:Real},
    ωs::AbstractMatrix{<:Real},
)

where αs are the weights of dimension (A, ), γs is the covariance matrix of dimension (D, A) and ωs are the mean vectors and is of dimension (D, A). Here, D is input dimension and A is the number of spectral components.

h is the kernel, which defaults to SqExponentialKernel if not specified.

Generalised Spectral Mixture kernel function. This family of functions is dense in the family of stationary real-valued kernels with respect to the pointwise convergence.[1]

\[ κ(x, y) = αs' (h(-(γs' * t)^2) .* cos(π * ωs' * t), t = x - y\]

References:

[1] Generalized Spectral Kernels, by Yves-Laurent Kom Samo and Stephen J. Roberts
[2] SM: Gaussian Process Kernels for Pattern Discovery and Extrapolation,
        ICML, 2013, by Andrew Gordon Wilson and Ryan Prescott Adams,
[3] Covariance kernels for fast automatic pattern discovery and extrapolation
    with Gaussian processes, Andrew Gordon Wilson, PhD Thesis, January 2014.
    http://www.cs.cmu.edu/~andrewgw/andrewgwthesis.pdf
[4] http://www.cs.cmu.edu/~andrewgw/pattern/.

KernelFunctions.spectral_mixture_product_kernel — Function

spectral_mixture_product_kernel(
    h::Kernel=SqExponentialKernel(),
    αs::AbstractMatrix{<:Real},
    γs::AbstractMatrix{<:Real},
    ωs::AbstractMatrix{<:Real},
)

where αs are the weights of dimension (D, A), γs is the covariance matrix of dimension (D, A) and ωs are the mean vectors and is of dimension (D, A). Here, D is input dimension and A is the number of spectral components.

Spectral Mixture Product Kernel. With enough components A, the SMP kernel can model any product kernel to arbitrary precision, and is flexible even with a small number of components [1]

h is the kernel, which defaults to SqExponentialKernel if not specified.

\[ κ(x, y) = Πᵢ₌₁ᴷ Σ(αsᵢᵀ .* (h(-(γsᵢᵀ * tᵢ)²) .* cos(ωsᵢᵀ * tᵢ))), tᵢ = xᵢ - yᵢ\]

References:

[1] GPatt: Fast Multidimensional Pattern Extrapolation with GPs,
    arXiv 1310.5288, 2013, by Andrew Gordon Wilson, Elad Gilboa,
    Arye Nehorai and John P. Cunningham

KernelFunctions.PeriodicKernel — Type

PeriodicKernel(r::AbstractVector)
PeriodicKernel(dims::Int)
PeriodicKernel(T::DataType, dims::Int)

Periodic Kernel as described in http://www.inference.org.uk/mackay/gpB.pdf eq. 47.

    κ(x,y) = exp( - 0.5 sum_i(sin (π(x_i - y_i))/r_i))

KernelFunctions.WienerKernel — Type

WienerKernel{i}()

i-times integrated Wiener process kernel function.

For i=-1, this is just the white noise covariance, see WhiteKernel.

For i= 0, this is the Wiener process covariance,

For i= 1, this is the integrated Wiener process covariance (velocity),

For i= 2, this is the twice-integrated Wiener process covariance (accel.),

For i= 3, this is the thrice-integrated Wiener process covariance,

where $κᵢ$ is given by

\[ κ₋₁(x, y) = δ(x, y) κ₀(x, y) = min(x, y)\]

and for $i >= 1$,

\[ κᵢ(x, y) = 1 / aᵢ * min(x, y)^(2i + 1) + bᵢ * min(x, y)^(i + 1) * |x - y| * rᵢ(x, y),\]

with the coefficients ``aᵢ``, ``bᵢ`` and the residual ``rᵢ(x, y)`` defined as follows:

\[ a₁ = 3, b₁ = 1/2, r₁(x, y) = 1, a₂ = 20, b₂ = 1/12, r₂(x, y) = x + y - min(x, y) / 2, a₃ = 252, b₃ = 1/720, r₃(x, y) = 5 * max(x, y)² + 2 * x * z + 3 * min(x, y)² \]

References:

See the paper Probabilistic ODE Solvers with Runge-Kutta Means by Schober, Duvenaud and Hennig, NIPS, 2014, for more details.

Composite Kernels

KernelFunctions.TransformedKernel — Type

TransformedKernel(k::Kernel,t::Transform)

Return a kernel where inputs are pretransformed by t : k(t(x),t(x')) Can also be called via transform : transform(k, t)

KernelFunctions.ScaledKernel — Type

ScaledKernel(k::Kernel, σ²::Real)

Return a kernel premultiplied by the variance σ² : σ² k(x,x')

KernelFunctions.KernelSum — Type

KernelSum <: Kernel

Create a sum of kernels. One can also use the operator +.

There are various ways in which you create a KernelSum:

The simplest way to specify a KernelSum would be to use the overloaded + operator. This is equivalent to creating a KernelSum by specifying the kernels as the arguments to the constructor.

julia> k1 = SqExponentialKernel(); k2 = LinearKernel(); X = rand(5);

julia> (k = k1 + k2) == KernelSum(k1, k2)
true

julia> kernelmatrix(k1 + k2, X) == kernelmatrix(k1, X) .+ kernelmatrix(k2, X)
true

julia> kernelmatrix(k, X) == kernelmatrix(k1 + k2, X)
true

You could also specify a KernelSum by providing a Tuple or a Vector of the kernels to be summed. We suggest you to use a Tuple when you have fewer components and a Vector when dealing with a large number of components.

julia> KernelSum((k1, k2)) == k1 + k2
true

julia> KernelSum([k1, k2]) == KernelSum((k1, k2)) == k1 + k2
true

KernelFunctions.KernelProduct — Type

KernelProduct <: Kernel

Create a product of kernels. One can also use the overloaded operator *.

There are various ways in which you create a KernelProduct:

The simplest way to specify a KernelProduct would be to use the overloaded * operator. This is equivalent to creating a KernelProduct by specifying the kernels as the arguments to the constructor.

julia> k1 = SqExponentialKernel(); k2 = LinearKernel(); X = rand(5);

julia> (k = k1 * k2) == KernelProduct(k1, k2)
true

julia> kernelmatrix(k1 * k2, X) == kernelmatrix(k1, X) .* kernelmatrix(k2, X)
true

julia> kernelmatrix(k, X) == kernelmatrix(k1 * k2, X)
true

You could also specify a KernelProduct by providing a Tuple or a Vector of the kernels to be multiplied. We suggest you to use a Tuple when you have fewer components and a Vector when dealing with a large number of components.

julia> KernelProduct((k1, k2)) == k1 * k2
true

julia> KernelProduct([k1, k2]) == KernelProduct((k1, k2)) == k1 * k2
true

KernelFunctions.TensorProduct — Type

TensorProduct(kernels...)

Create a tensor product kernel from kernels $k_1, \ldots, k_n$, i.e., a kernel $k$ that is given by

\[k(x, y) = \prod_{i=1}^n k_i(x_i, y_i).\]

The kernels can be specified as individual arguments, a tuple, or an iterable data structure such as an array. Using a tuple or individual arguments guarantees that TensorProduct is concretely typed but might lead to large compilation times if the number of kernels is large.

Transforms

KernelFunctions.Transform — Type

Abstract type defining a slice-wise transformation on an input matrix

KernelFunctions.IdentityTransform — Type

IdentityTransform()

Return exactly the input

KernelFunctions.ScaleTransform — Type

ScaleTransform(l::Real)

Multiply every element of the input by l

    l = 2.0
    tr = ScaleTransform(l)

KernelFunctions.ARDTransform — Type

ARDTransform(v::AbstractVector)
ARDTransform(s::Real, dims::Int)

Multiply every vector of observation by v element-wise

    v = rand(3)
    tr = ARDTransform(v)

KernelFunctions.LinearTransform — Type

LinearTransform(A::AbstractMatrix)

Apply the linear transformation realised by the matrix A.

The second dimension of A must match the number of features of the target.

Examples

julia> A = rand(10, 5)
julia> tr = LinearTransform(A)

KernelFunctions.FunctionTransform — Type

FunctionTransform(f)

Take a function or object f as an argument which is going to act on each vector individually. Make sure that f is supposed to act on a vector. For example replace f(x)=sin(x) by f(x)=sin.(x)

    f(x) = abs.(x)
    tr = FunctionTransform(f)

KernelFunctions.SelectTransform — Type

SelectTransform(dims)

Select the dimensions dims that the kernel is applied to.

    dims = [1,3,5,6,7]
    tr = SelectTransform(dims)
    X = rand(100,10)
    transform(tr,X,obsdim=2) == X[dims,:]

KernelFunctions.ChainTransform — Type

ChainTransform(ts::AbstractVector{<:Transform})

Chain a series of transform, here t1 will be called first

    t1 = ScaleTransform()
    t2 = LinearTransform(rand(3,4))
    ct = ChainTransform([t1,t2]) #t1 will be called first
    ct == t2 ∘ t1

Functions

KernelFunctions.kernelmatrix — Function

kernelmatrix(κ::Kernel, X; obsdim::Int = 2)
kernelmatrix(κ::Kernel, X, Y; obsdim::Int = 2)

Calculate the kernel matrix of X (and Y) with respect to kernel κ. obsdim = 1 means the matrix X (and Y) has size #samples x #dimension obsdim = 2 means the matrix X (and Y) has size #dimension x #samples

KernelFunctions.kernelmatrix! — Function

kernelmatrix!(K::AbstractMatrix, κ::Kernel, X; obsdim::Integer = 2)
kernelmatrix!(K::AbstractMatrix, κ::Kernel, X, Y; obsdim::Integer = 2)

In-place version of kernelmatrix where pre-allocated matrix K will be overwritten with the kernel matrix.

KernelFunctions.kerneldiagmatrix — Function

kerneldiagmatrix(κ::Kernel, X; obsdim::Int = 2)

Calculate the diagonal matrix of X with respect to kernel κ obsdim = 1 means the matrix X has size #samples x #dimension obsdim = 2 means the matrix X has size #dimension x #samples

kerneldiagmatrix(κ::Kernel, X, Y; obsdim::Int = 2)

Calculate the diagonal of kernelmatrix(κ, X, Y; obsdim) efficiently. Requires that X and Y are the same length.

KernelFunctions.kerneldiagmatrix! — Function

kerneldiagmatrix!(K::AbstractVector, κ::Kernel, X; obsdim::Int = 2)
kerneldiagmatrix!(K::AbstractVector, κ::Kernel, X, Y; obsdim::Int = 2)

In place version of kerneldiagmatrix

Missing docstring.

Missing docstring for kernelpdmat. Check Documenter's build log for details.

Missing docstring.

Missing docstring for kernelkronmat. Check Documenter's build log for details.

KernelFunctions.nystrom — Function

nystrom(k::Kernel, X::Matrix, S::Vector; obsdim::Int=defaultobs)

Computes a factorization of Nystrom approximation of the square kernel matrix of data matrix X with respect to kernel k. Returns a NystromFact struct which stores a Nystrom factorization satisfying:

\[\mathbf{K} \approx \mathbf{C}^{\intercal}\mathbf{W}\mathbf{C}\]

nystrom(k::Kernel, X::Matrix, r::Real; obsdim::Int=defaultobs)

Computes a factorization of Nystrom approximation of the square kernel matrix of data matrix X with respect to kernel k using a sample ratio of r. Returns a NystromFact struct which stores a Nystrom factorization satisfying:

\[\mathbf{K} \approx \mathbf{C}^{\intercal}\mathbf{W}\mathbf{C}\]

KernelFunctions.transform — Function

    transform(k::Kernel, t::Transform) (1)
    transform(k::Kernel, ρ::Real) (2)
    transform(k::Kernel, ρ::AbstractVector) (3)

(1) Create a TransformedKernel with transform t and kernel k (2) Same as (1) with a ScaleTransform with scale ρ (3) Same as (1) with an ARDTransform with scales ρ

Utilities

KernelFunctions.ColVecs — Type

ColVecs(X::AbstractMatrix)

A lightweight wrapper for an AbstractMatrix to make it behave like a vector of vectors. Each vector represents a column of the matrix

KernelFunctions.RowVecs — Type

RowVecs(X::AbstractMatrix)

A lightweight wrapper for an AbstractMatrix to make it behave like a vector of vectors. Each vector represents a row of the matrix

KernelFunctions.NystromFact — Type

NystromFact

Type for storing a Nystrom factorization. The factorization contains two fields: W and C, two matrices satisfying:

\[\mathbf{K} \approx \mathbf{C}^{\intercal}\mathbf{W}\mathbf{C}\]

Index

KernelFunctions.ARDTransform

KernelFunctions.ChainTransform

KernelFunctions.ColVecs

KernelFunctions.ConstantKernel

KernelFunctions.CosineKernel

KernelFunctions.ExponentialKernel

KernelFunctions.ExponentiatedKernel

KernelFunctions.EyeKernel

KernelFunctions.FBMKernel

KernelFunctions.FunctionTransform

KernelFunctions.GaborKernel

KernelFunctions.GammaExponentialKernel

KernelFunctions.GammaRationalQuadraticKernel

KernelFunctions.IdentityTransform

KernelFunctions.KernelProduct

KernelFunctions.KernelSum

KernelFunctions.LinearKernel

KernelFunctions.LinearTransform

KernelFunctions.MahalanobisKernel

KernelFunctions.Matern32Kernel

KernelFunctions.Matern52Kernel

KernelFunctions.MaternKernel

KernelFunctions.NeuralNetworkKernel

KernelFunctions.NystromFact

KernelFunctions.PeriodicKernel

KernelFunctions.PiecewisePolynomialKernel

KernelFunctions.PolynomialKernel

KernelFunctions.RationalQuadraticKernel

KernelFunctions.RowVecs

KernelFunctions.ScaleTransform

KernelFunctions.ScaledKernel

KernelFunctions.SelectTransform

KernelFunctions.SqExponentialKernel

KernelFunctions.TensorProduct

KernelFunctions.Transform

KernelFunctions.TransformedKernel

KernelFunctions.WhiteKernel

KernelFunctions.WienerKernel

KernelFunctions.ZeroKernel

KernelFunctions.kerneldiagmatrix

KernelFunctions.kerneldiagmatrix!

KernelFunctions.kernelmatrix

KernelFunctions.kernelmatrix!

KernelFunctions.nystrom

KernelFunctions.spectral_mixture_kernel

KernelFunctions.spectral_mixture_product_kernel

KernelFunctions.transform