API Library

API Library

Module

Kernel Functions

Kernel Combinations

Transforms

Functions

Index

Module

Missing docstring.

Missing docstring for KernelFunctions. Check Documenter's build log for details.

Kernel Functions

KernelFunctions.SqExponentialKernel — Type

SqExponentialKernel([ρ=1.0])

The squared exponential kernel is an isotropic Mercer kernel given by the formula:

    κ(x,y) = exp(-ρ²‖x-y‖²)

See also ExponentialKernel for a related form of the kernel or GammaExponentialKernel for a generalization.

KernelFunctions.ExponentialKernel — Type

ExponentialKernel([ρ=1.0]) The exponential kernel is an isotropic Mercer kernel given by the formula:

    κ(x,y) = exp(-ρ‖x-y‖)

KernelFunctions.GammaExponentialKernel — Type

GammaExponentialKernel([ρ=1.0,[gamma=2.0]]) The γ-exponential kernel is an isotropic Mercer kernel given by the formula:

    κ(x,y) = exp(-ρ^(2γ)‖x-y‖^(2γ))

KernelFunctions.ExponentiatedKernel — Type

ExponentiatedKernel([ρ=1]) The exponentiated kernel is a Mercer kernel given by:

    κ(x,y) = exp(ρ²xᵀy)

KernelFunctions.MaternKernel — Type

MaternKernel([ρ=1.0,[ν=1.0]]) The matern kernel is an isotropic Mercer kernel given by the formula:

    κ(x,y) = 2^{1-ν}/Γ(ν)*(√(2ν)‖x-y‖)^ν K_ν(√(2ν)‖x-y‖)

For ν=n+1/2, n=0,1,2,... it can be simplified and you should instead use ExponentialKernel for n=0, Matern32Kernel, for n=1, Matern52Kernel for n=2 and SqExponentialKernel for n=∞.

KernelFunctions.Matern32Kernel — Type

Matern32Kernel([ρ=1.0]) The matern 3/2 kernel is an isotropic Mercer kernel given by the formula:

    κ(x,y) = (1+√(3)ρ‖x-y‖)exp(-√(3)ρ‖x-y‖)

KernelFunctions.Matern52Kernel — Type

Matern52Kernel([ρ=1.0]) The matern 5/2 kernel is an isotropic Mercer kernel given by the formula:

    κ(x,y) = (1+√(5)ρ‖x-y‖ + 5ρ²‖x-y‖^2/3)exp(-√(5)ρ‖x-y‖)

KernelFunctions.LinearKernel — Type

LinearKernel([ρ=1.0,[c=0.0]]) The linear kernel is a Mercer kernel given by

    κ(x,y) = ρ²xᵀy + c

Where c is a real number

KernelFunctions.PolynomialKernel — Type

PolynomialKernel([ρ=1.0[,d=2.0[,c=0.0]]]) The polynomial kernel is a Mercer kernel given by

    κ(x,y) = (ρ²xᵀy + c)^d

Where c is a real number, and d is a shape parameter bigger than 1

KernelFunctions.RationalQuadraticKernel — Type

RationalQuadraticKernel([ρ=1.0[,α=2.0]]) The rational-quadratic kernel is an isotropic Mercer kernel given by the formula:

    κ(x,y)=(1+ρ²||x−y||²/α)^(-α)

where α is a shape parameter of the Euclidean distance. Check GammaRationalQuadraticKernel for a generalization.

KernelFunctions.GammaRationalQuadraticKernel — Type

GammaRationalQuadraticKernel([ρ=1.0[,α=2.0[,γ=2.0]]]) The Gamma-rational-quadratic kernel is an isotropic Mercer kernel given by the formula:

    κ(x,y)=(1+ρ^(2γ)||x−y||^(2γ)/α)^(-α)

where α is a shape parameter of the Euclidean distance and γ is another shape parameter.

KernelFunctions.ZeroKernel — Type

ZeroKernel([tr=IdentityTransform()])

Create a kernel always returning zero

KernelFunctions.ConstantKernel — Type

ConstantKernel([tr=IdentityTransform(),[c=1.0]])

    κ(x,y) = c

Kernel function always returning a constant value c

KernelFunctions.WhiteKernel — Type

WhiteKernel([tr=IdentityTransform()])

    κ(x,y) = δ(x,y)

Kernel function working as an equivalent to add white noise.

Kernel Combinations

KernelFunctions.KernelSum — Type

KernelSum(kernels::Array{Kernel};weights::Array{Real}=ones(length(kernels))) Create a positive weighted sum of kernels. One can also use the operator +

k1 = SqExponentialKernel()
k2 = LinearKernel()
k = KernelSum([k1,k2])
kernelmatrix(k,X) == kernelmatrix(k1,X).+kernelmatrix(k2,X)
kernelmatrix(k,X) == kernelmatrix(k1+k2,X)
kweighted = 0.5*k1 + 2.0*k2

KernelFunctions.KernelProduct — Type

KernelProduct(kernels::Array{Kernel}) Create a multiplication of kernels. One can also use the operator *

k1 = SqExponentialKernel()
k2 = LinearKernel()
k = KernelProduct([k1,k2])
kernelmatrix(k,X) == kernelmatrix(k1,X).*kernelmatrix(k2,X)
kernelmatrix(k,X) == kernelmatrix(k1*k2,X)

Transforms

KernelFunctions.Transform — Type

Abstract type defining a slice-wise transformation on an input matrix

KernelFunctions.IdentityTransform — Type

IdentityTransform Return exactly the input

KernelFunctions.ScaleTransform — Type

Scale Transform

    l = 2.0
    tr = ScaleTransform(l)

Multiply every element of the input by l

KernelFunctions.LowRankTransform — Type

LowRankTransform

    P = rand(10,5)
    tr = LowRankTransform(P)

Apply the low-rank projection realised by the matrix P The second dimension of P must match the number of features of the target.

KernelFunctions.FunctionTransform — Type

FunctionTransform

    f(x) = abs.(x)
    tr = FunctionTransform(f)

Take a function f as an argument which is going to act on each vector individually. Make sure that f is supposed to act on a vector by eventually using broadcasting For example f(x)=sin(x) -> f(x)=sin.(x)

KernelFunctions.ChainTransform — Type

Chain a series of transform, here t1 will be called first

    t1 = ScaleTransform()
    t2 = LowRankTransform(rand(3,4))
    ct = ChainTransform([t1,t2]) #t1 will be called first
    ct == t2∘t1

Functions

KernelFunctions.kernelmatrix — Function

    kernelmatrix(κ::Kernel, X::Matrix ; obsdim::Int=2)
    kernelmatrix(κ::Kernel, X::Matrix, Y::Matrix; obsdim::Int=2)

Calculate the kernel matrix of X (and Y) with respect to kernel κ. obsdim=1 means the matrix X (and Y) has size #samples x #dimension obsdim=2 means the matrix X (and Y) has size #dimension x #samples

KernelFunctions.kernelmatrix! — Function

    kernelmatrix!(K::Matrix, κ::Kernel, X::Matrix; obsdim::Integer=2)
    kernelmatrix!(K::Matrix, κ::Kernel, X::Matrix, Y::Matrix; obsdim::Integer=2)

In-place version of kernelmatrix where pre-allocated matrix K will be overwritten with the kernel matrix.

KernelFunctions.kerneldiagmatrix — Function

    kerneldiagmatrix(κ::Kernel, X::Matrix; obsdim::Int=2)

Calculate the diagonal matrix of X with respect to kernel κ obsdim=1 means the matrix X has size #samples x #dimension obsdim=2 means the matrix X has size #dimension x #samples

KernelFunctions.kerneldiagmatrix! — Function

    kerneldiagmatrix!(K::AbstractVector,κ::Kernel, X::Matrix; obsdim::Int=2)

In place version of kerneldiagmatrix

KernelFunctions.kernelpdmat — Function

Compute a positive-definite matrix in the form of a `PDMat` matrix see [PDMats.jl]() with the cholesky decomposition precomputed
The algorithm recursively tries to add recursively a diagonal nugget until positive definiteness is achieved or that the noise is too big

KernelFunctions.transform — Function

    transform(t::Transform, X::AbstractMatrix)
    transform(k::Kernel, X::AbstractMatrix)

Apply the transfomration t or k.transform on the input X

Index

KernelFunctions.ChainTransform

KernelFunctions.ConstantKernel

KernelFunctions.ExponentialKernel

KernelFunctions.ExponentiatedKernel

KernelFunctions.FunctionTransform

KernelFunctions.GammaExponentialKernel

KernelFunctions.GammaRationalQuadraticKernel

KernelFunctions.IdentityTransform

KernelFunctions.KernelProduct

KernelFunctions.KernelSum

KernelFunctions.LinearKernel

KernelFunctions.LowRankTransform

KernelFunctions.Matern32Kernel

KernelFunctions.Matern52Kernel

KernelFunctions.MaternKernel

KernelFunctions.PolynomialKernel

KernelFunctions.RationalQuadraticKernel

KernelFunctions.ScaleTransform

KernelFunctions.SqExponentialKernel

KernelFunctions.Transform

KernelFunctions.WhiteKernel

KernelFunctions.ZeroKernel

KernelFunctions.kerneldiagmatrix

KernelFunctions.kerneldiagmatrix!

KernelFunctions.kernelmatrix

KernelFunctions.kernelmatrix!

KernelFunctions.kernelpdmat

KernelFunctions.transform