API

API Library

API Library

Module

Kernel Functions

Kernel Combinations

Transforms

Functions

Index

Module

Missing docstring.

Missing docstring for KernelFunctions. Check Documenter's build log for details.

Kernel Functions

KernelFunctions.SqExponentialKernel — Type.

SqExponentialKernel([ρ=1.0])

The squared exponential kernel is an isotropic Mercer kernel given by the formula:

    κ(x,y) = exp(-‖x-y‖²)

See also ExponentialKernel for a related form of the kernel or GammaExponentialKernel for a generalization.

source

Missing docstring.

Missing docstring for Exponential. Check Documenter's build log for details.

KernelFunctions.GammaExponentialKernel — Type.

GammaExponentialKernel([ρ=1.0,[gamma=2.0]])

The γ-exponential kernel is an isotropic Mercer kernel given by the formula:

    κ(x,y) = exp(-‖x-y‖^2γ)

Examples

julia> GammaExponentialKernel()
GammaExponentialKernel{Float64,Float64,Float64}(1.0,2.0)

julia> GammaExponentialKernel(2.0f0,3.0)
GammaExponentialKernel{Float32,Float32,Float64}(2.0,3.0)

julia> GammaExponentialKernel([2.0,3.0],2f0)
GammaExponentialKernel{Float64,Array{Float64},Float32}([2.0,3.0],2.0)

source

KernelFunctions.ExponentiatedKernel — Type.

ExponentiatedKernel([α=1])

The exponentiated kernel is a Mercer kernel given by:

        κ(x,y) = exp(xᵀy)

source

KernelFunctions.MaternKernel — Type.

MaternKernel([ρ=1.0,[ν=1.0]])

The matern kernel is an isotropic Mercer kernel given by the formula:

    κ(x,y) = 2^{1-ν}/Γ(ν)*(√(2ν)‖x-y‖)^ν K_ν(√(2ν)‖x-y‖)

For ν=n+1/2, n=0,1,2,... it can be simplified and you should instead use ExponentialKernel for n=0, Matern32Kernel, for n=1, Matern52Kernel for n=2 and SqExponentialKernel for n=∞. ρ is the lengthscale parameter(s) or the transform object.

source

KernelFunctions.Matern32Kernel — Type.

Matern32Kernel(ρ=1.0)

The matern 3/2 kernel is an isotropic Mercer kernel given by the formula:

    κ(x,y) = (1+√(3)‖x-y‖)exp(√(3)‖x-y‖)

ρ is the lengthscale parameter(s) or a transform object.

source

KernelFunctions.Matern52Kernel — Type.

Matern52Kernel(ρ=1.0)

The matern 5/2 kernel is an isotropic Mercer kernel given by the formula:

    κ(x,y) = (1+√(5)‖x-y‖ + 5‖x-y‖^2/3)exp(√(5)‖x-y‖)

ρ is the lengthscale parameter(s) or a transform object.

source

KernelFunctions.LinearKernel — Type.

LinearKernel([ρ=1.0,[c=0.0]])

The linear kernel is a Mercer kernel given by

    κ(x,y) = xᵀy + c

Where `c` is a real number

source

KernelFunctions.PolynomialKernel — Type.

PolynomialKernel([ρ=1.0[,d=2.0[,c=0.0]]])

The polynomial kernel is a Mercer kernel given by

    κ(x,y) = (xᵀy + c)^d

Where `c` is a real number, and `d` is a shape parameter bigger than 1

source

KernelFunctions.RationalQuadraticKernel — Type.

RationalQuadraticKernel([ρ=1.0[,α=2.0]])

The rational-quadratic kernel is an isotropic Mercer kernel given by the formula:

    κ(x,y)=(1+||x−y||^2/α)^(-α)

where `α` is a shape parameter of the Euclidean distance. Check `GammaRationalQuadraticKernel` for a generalization.

source

KernelFunctions.GammaRationalQuadraticKernel — Type.

GammaRationalQuadraticKernel([ρ=1.0[,α=2.0[,γ=2.0]]])

The Gamma-rational-quadratic kernel is an isotropic Mercer kernel given by the formula:

    κ(x,y)=(1+||x−y||^(2γ)/α)^(-α)

where α is a shape parameter of the Euclidean distance and γ is another shape parameter.

source

Kernel Combinations

Missing docstring.

Missing docstring for KernelSum. Check Documenter's build log for details.

Missing docstring.

Missing docstring for KernelProduct. Check Documenter's build log for details.

Transforms

Missing docstring.

Missing docstring for Transform. Check Documenter's build log for details.

Missing docstring.

Missing docstring for IdentityTransform. Check Documenter's build log for details.

KernelFunctions.ScaleTransform — Type.

Scale Transform

source

Missing docstring.

Missing docstring for LowRankTransform. Check Documenter's build log for details.

KernelFunctions.FunctionTransform — Type.

FunctionTransform

Take a function `f` as an argument which is going to act on each vector individually.
Make sure that `f` is supposed to act on a vector by eventually using broadcasting
For example `f(x)=sin(x)` -> `f(x)=sin.(x)`

source

Functions

KernelFunctions.kernelmatrix — Function.

    kernelmatrix(κ::Kernel, X::Matrix ; obsdim::Int=2, symmetrize::Bool=true)

Calculate the kernel matrix of X with respect to kernel κ. obsdim=1 means the matrix X has size #samples x #dimension obsdim=2 means the matrix X has size #dimension x #samples

source

    kernelmatrix(κ::Kernel, X::Matrix, Y::Matrix; obsdim::Int=2)

Calculate the base matrix of X and Y with respect to kernel κ. obsdim=1 means the matrices X and Y have sizes #samples x #dimension obsdim=2 means the matrices X and Y have size #dimension x #samples

source

KernelFunctions.kernelmatrix! — Function.

    kernelmatrix!(K::Matrix, κ::Kernel, X::Matrix; obsdim::Integer=2, symmetrize::Bool=true)

In-place version of kernelmatrix where pre-allocated matrix K will be overwritten with the kernel matrix.

source

    kernelmatrix!(K::Matrix, κ::Kernel, X::Matrix, Y::Matrix; obsdim::Integer=2)

In-place version of kernelmatrix where pre-allocated matrix K will be overwritten with the kernel matrix.

source

KernelFunctions.kerneldiagmatrix — Function.

    kerneldiagmatrix(κ::Kernel, X::Matrix; obsdim::Int=2)

Calculate the diagonal matrix of X with respect to kernel κ obsdim=1 means the matrix X has size #samples x #dimension obsdim=2 means the matrix X has size #dimension x #samples

source

Missing docstring.

Missing docstring for kerneldiagmatrix!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for transform. Check Documenter's build log for details.

Index

KernelFunctions.ExponentiatedKernel

KernelFunctions.FunctionTransform

KernelFunctions.GammaExponentialKernel

KernelFunctions.GammaRationalQuadraticKernel

KernelFunctions.LinearKernel

KernelFunctions.Matern32Kernel

KernelFunctions.Matern52Kernel

KernelFunctions.MaternKernel

KernelFunctions.PolynomialKernel

KernelFunctions.RationalQuadraticKernel

KernelFunctions.ScaleTransform

KernelFunctions.SqExponentialKernel

KernelFunctions.kerneldiagmatrix

KernelFunctions.kernelmatrix

KernelFunctions.kernelmatrix!