Kernel Functions

Exponential Kernels

KernelFunctions.ExponentialKernel — Type.

ExponentialKernel([ρ=1.0])

The exponential kernel is an isotropic Mercer kernel given by the formula:

    κ(x,y) = exp(-‖x-y‖)

source

Missing docstring.

Missing docstring for SqExponentialKernel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GammaExponentialKernel. Check Documenter's build log for details.

Matern Kernels

Missing docstring.

Missing docstring for MaternKernel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Matern32Kernel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Matern52Kernel. Check Documenter's build log for details.

Polynomial Kernels

Missing docstring.

Missing docstring for LinearKernel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PolynomialKernel. Check Documenter's build log for details.

Constant Kernels

KernelFunctions.ConstantKernel — Type.

ConstantKernel([c=1.0])

    κ(x,y) = c

Kernel function always returning a constant value `c`

source

KernelFunctions.WhiteKernel — Type.

WhiteKernel()

    κ(x,y) = δ(x,y)

Kernel function working as an equivalent to add white noise.

source

KernelFunctions.ZeroKernel — Type.

ZeroKernel()

Create a kernel that always return a zero kernel matrix

source